// webserver/src/dtos/response/types.ts
export interface Item {
  id: string;
  title: string;
  price: Price;
  picture: string;
  condition: string;
  free_shipping: boolean;
  state?: string;
}

export interface Price {
  currency: string;
  amount: number;
  decimals: number;
}

export interface ResponseAuthor {
  author: Author;
}

export interface ItemsContainer extends ResponseAuthor {
  categories: string[];
  items: Item[];
}

export interface Author {
  name: string;
  lastname: string;
}

export interface ItemDetail extends Item {
  sold_quantity: number;
  description: string;
  categories: string[];
}

export interface ItemDetailContainer extends ResponseAuthor {
  item: ItemDetail;
}